Allow the Bridge code to display error messages to the user
In some cases, it can be useful for the Bridge code to return an explicit error message to the user in case of authentication failure. Currently, MushRaider only says “The email or password that you entered is incorrect”. This can now be accomplished by returning an ‘error’ field in the JSON output that contains the error message to display.

Of course, it is the Bridge developer’s responsibility to send something meaningful and not expose any secret or inner working of the Bridge code or the server settings.

If the ‘error’ field is not returned, the original MushRaider message will be shown.